feat: rework presets panel to use collapsible panels instead of an extra page of buttons
